block: simplify issueing discard, write_same, zeroout
Simplify special cases for issueing discard, write_same, and zeroout,
replacing bio_batch completions with submit_bio_wait(). This conversion
is possible because generic_make_request() will now do for us what the
code in blk-lib.c was doing manually, with the bio_batch stuff. So we
still need some looping in case we're trying to discard/zeroout more
than around a gigabyte, but when we can submit that much at a time
doing the submissions in parallel really shouldn't matter.

Signed-off-by: Kent Overstreet <kmo@daterainc.com>
[dpark: add more description in commit message]
Signed-off-by: Dongsu Park <dongsu.park@profitbricks.com>
Cc: Jens Axboe <axboe@kernel.dk>