nfsd: ensure that nfs4_file_get_access enforces deny modes
The current enforcement of deny modes is both inefficient and scattered.
The former is a problem now, and the latter problem will mean races once
the client_mutex is removed.

First, we address the inefficiency. We (necessarily) track deny modes on
a per-stateid basis, but when we go to enforce them we have to walk the
entire list of stateids and check against each one. Instead of doing
that, maintain a per-nfs4_file deny mode.

When a file is opened, we simply set any deny bits in that mode that
were specified in the OPEN call. We can then use that unified deny mode
to do a simple check to see whether there are any conflicts without
needing to walk the entire stateid list.

The only time we'll need to walk the entire list of stateids is when
a stateid that has a deny mode on it is being released, or one is having
its deny mode downgraded. In that case, we must walk the entire list
and recalculate the fi_share_deny field. Since deny modes are pretty
rare today, this shouldn't happen much under normal workloads.

To address the potential for races once the client_mutex is removed, we
first check for conflicting deny modes in nfs4_file_get_access prior to
opening the file. If it turns out that we need to do a VFS layer open of
the file, then do so and recheck for conflicting deny modes afterward.
If there are any, then just put access to the file and return
nfserr_share_denied.

Finally, deal with potential races in get_lock_access. Taking an
fi_access reference must be done under the fi_lock, or that could race
with a nfs4_file_put_access call. Since we will have just dropped the
lock after finding a readable or writeable file, add some *_locked
variants of find_readable_file and find_writeable_file that we can call
while already holding the fi_lock.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>